--- src/advanced/src/components/CartDisplay/AddToCartForm.tsx
@@ -1,115 +1,0 @@
-import { useRef, useState, useEffect } from 'react';
-import { useCart } from '../../hooks/useCart';
-import {
-  getProducts,
-  setProductUpdateCallback,
-  setSelectedProduct,
-  type Product,
-} from '../../services/saleService';
-
-export default function AddToCartForm() {
-  const { dispatch } = useCart();
-  const selectRef = useRef<HTMLSelectElement>(null);
-  const [products, setProducts] = useState<Product[]>(getProducts());
-
-  useEffect(() => {
-    setProductUpdateCallback((updatedProducts) => {
-      setProducts([...updatedProducts]);
-    });
-  }, []);
-
-  const handleAddToCart = () => {
-    const selected = selectRef.current?.value;
-    if (!selected) return;
-
-    const selectedProduct = products.find((p) => p.id === selected);
-    if (selectedProduct && selectedProduct.quantity === 0) {
-      alert('재고가 부족합니다.');
-      return;
-    }
-
-    dispatch({
-      type: 'ADD_ITEM',
-      payload: {
-        productId: selected,
-        quantity: 1,
-      },
-    });
-  };
-
-  const formatProductOption = (product: Product) => {
-    const hasDiscount = product.price < product.originalPrice;
-    const discountRate = hasDiscount
-      ? Math.round((1 - product.price / product.originalPrice) * 100)
-      : 0;
-
-    if (product.quantity === 0) {
-      return `${product.name} - ₩${product.originalPrice.toLocaleString()} (품절)`;
-    }
-
-    if (hasDiscount) {
-      return `${product.saleIcon}${product.name} - ₩${product.originalPrice.toLocaleString()} → ₩${product.price.toLocaleString()} (${discountRate}% SALE!)`;
-    }
-
-    return `${product.name} - ₩${product.price.toLocaleString()}`;
-  };
-
-  const handleSelectChange = () => {
-    const selected = selectRef.current?.value;
-    if (selected) {
-      setSelectedProduct(selected);
-    }
-  };
-
-  return (
-    <div className="mb-6 pb-6 border-b border-gray-200">
-      <select
-        id="product-select"
-        className="w-full p-3 border border-gray-300 rounded-lg text-base mb-3"
-        ref={selectRef}
-        onChange={handleSelectChange}
-      >
-        {products.map((product) => (
-          <option
-            key={product.id}
-            value={product.id}
-            disabled={product.quantity === 0}
-            className={
-              product.quantity === 0
-                ? 'text-gray-400'
-                : product.price < product.originalPrice
-                  ? 'text-red-500 font-bold'
-                  : ''
-            }
-          >
-            {formatProductOption(product)}
-          </option>
-        ))}
-      </select>
-      <button
-        id="add-to-cart"
-        className="w-full py-3 bg-black text-white text-sm font-medium uppercase tracking-wider hover:bg-gray-800 transition-all"
-        onClick={handleAddToCart}
-      >
-        Add to Cart
-      </button>
-      {products.some((p) => p.quantity <= 5) && (
-        <div
-          id="stock-status"
-          className="text-xs text-red-500 mt-3 whitespace-pre-line"
-        >
-          {products
-            .filter((p) => p.quantity <= 5)
-            .map((p) => {
-              if (p.quantity === 0) {
-                return `${p.name}: 품절`;
-              } else {
-                return `${p.name}: 재고 부족 (${p.quantity}개 남음)`;
-              }
-            })
-            .join('\n')}
-        </div>
-      )}
-    </div>
-  );
-}
